+++ src/routes/(chat)/(chatinput)/chat/+page.ts
