+++ src/routes/(chat)/(chatinput)/chat/[chatId=uuid]/+page.server.ts
